requirements.txt
openai-whisper
ollama
opencv-python
gradio
xhtml2pdf